Fixed method

+++ src/main/java/org/piki/App.java
@@ -3,17 +3,18 @@ package org.piki;
 public class App {
     public static void main(String[] args) {
         System.out.println("Here are some Fibonacci numbers:");
-        for (int i=1; i<=10; i++) {
+        for (int i = 1; i <= 10; i++) {
             System.out.println(i + ": " + fibonacci(i));
         }
     }
 
-    public static int fibonacci(int n) {
-        if (n <= 2) return 1;
+    public static long fibonacci(int n) {
+        if (n <= 2)
+            return 1;
 
-        int a = 1, b = 1;
-        for (int i=2; i<n; i++) {
-            int t = a;
+        long a = 1, b = 1;
+        for (long i = 2; i < n; i++) {
+            long t = a;
             a = b;
             b = b + t;
         }
